requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
audioread==2.1.8
Automat==0.8.0
beautifulsoup4==4.9.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
Click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
deprecation==2.1.0
discogs-client==2.3.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.1
django-bootstrap4==2.2.0
django-composite-foreignkey==1.1.0
django-environ==0.4.5
django-extensions==3.0.9
entrypoints==0.3
eyeD3==0.9.5
ffmpeg==1.4
filelock==3.0.12
filetype==1.0.7
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.2
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.4
psycopg2-binary==2.8.6
pyacoustid==1.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDejavu==0.1.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
virtualenv==20.0.33
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
